Reformat: pybind11 is no longer included as a submodule
While having pybind11 as a submodule can be circumstantially
convenient, submodules in general are tempermental. Specifically,
when pulling or cloning the repository, the submodule's files will
not be added, but the submodule is still considered present, so
manually adding the submodule again doesn't work either. Going
forward, developers are expected to clone the the pybind11
repository separately and to include the library with CMake using
the BuildScript.py <WIP> file being added.