Refactored: removed unecessary options after changes to how pybind11 was included
Now that pybind11 is once again being used as a submodule, this CMakeLists.txt
file will follow the format of official examples and not the limited expertise
of the original committer. The script removes the build options, since the build
environment for the module still generates a VS solution file and doesn't need
a separate build mode.