Fixed: Removed cast function in lambda used to expose classicalBeam
As a stop gap measure, the classicalBeam function's return of a
std::vector is exposed directly and is implicitly cast as a python
list. This requires that GUI.py use the return of classicalBeam
to initialize a numpy array. Eventually, the casting option will
be reimplemented so that the functions returns a python ndarray.