Cut counter for hours wasted on the IDE not recognizing a macro
In the first attempts at constructing the build environment with cmake, the
Visual Studio IDE failed to recognize the PYBIND11_MODULE macro. The reason
for this was that the CmakeLists.txt file configured cmake to build an
executable rather than a downloadable library. There is no longer a purpose
for the counter or the sarcastic comment accompanying it now that the bug
has long been fixed.