Cut classical beam function
Since a working downloadable library can now be built using
the C++ code which contains a function for generating the
trajectory curve of the electron beam, the classical beam
function in this file is no longer needed. The analogous
function for the relativistic model has been reformatted
to conform with PEP standards and a TODO section in its
docstring now explains what needs to be done with the
function going forward.