Replaced function calculating beam trajectory
The previous function ClassicalBeam incorrectly calculated the trajectory of a charged particle in an external EM field. ClassicalBeam has been replaced with classical_beam which solves a coupled system of equations of motion (EOM) using scipy.integrate.odeint to derive the correct trajectory. The EOM that classical_beam uses is left as an argument and thus in the future, any function that accepts the required parameters and returns an array for the coupled system of EOMs can be passed to the classical_beam and solved numerically, in the event that more complicated simulations of different EM fields are added in the future.

There is currently one function that returns the system of EOMs, uniform_fields_eom, which returns the EOMs for the Lorentz force assuming a uniform EM field whose component magnitudes are given as parameters to be passed to the function.

The relativistic beam calculating function has been removed for the time being since continued work in the near future will focus on the classical beam calculation.

Global variables added for the gap between electrodes in the electron gun and the time interval for which the trajectory of the electrons is to be simulated.

Lastly, the source code was refactored to abide by PEP guidelines.